fixing a problem if you are using --fast-math compiler switch

+++ mpglib/interface.c
@@ -1,4 +1,4 @@
-/* $Id: interface.c,v 1.50 2007/07/01 20:05:52 robert Exp $ */
+/* $Id: interface.c,v 1.51 2008/04/22 23:01:22 robert Exp $ */
 
 #ifdef HAVE_CONFIG_H
 # include <config.h>
@@ -590,7 +590,7 @@ int decodeMP3_unclipped( PMPSTR mp,unsigned char *in,int isize,char *out,
                           int osize,int *done)
 {
         /* we forbid input with more than 1152 samples per channel for output in unclipped mode */
-        if(osize < 1152 * 2 * sizeof(real) ) {
+        if(osize < (int)(1152 * 2 * sizeof(real)) ) {
                 fprintf(stderr,"To less out space\n");
                 return MP3_ERR;
         }
